# platform/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogOriginView": {
        "properties": {
          "artifact_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Entry Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "origin_key": {
            "title": "Origin Key",
            "type": "string"
          },
          "origin_type": {
            "const": "upload",
            "title": "Origin Type",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "source_workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Workspace Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "id",
          "origin_type",
          "origin_key",
          "source_workspace_id"
        ],
        "title": "CatalogOriginView",
        "type": "object"
      },
      "CatalogPairOrigins": {
        "properties": {
          "architecture": {
            "title": "Architecture",
            "type": "string"
          },
          "code_id": {
            "title": "Code Id",
            "type": "string"
          },
          "debug_id": {
            "title": "Debug Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogOriginView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "pair_id": {
            "title": "Pair Id",
            "type": "string"
          },
          "qualification_version": {
            "title": "Qualification Version",
            "type": "integer"
          },
          "state": {
            "enum": [
              "active",
              "withdrawn"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "pair_id",
          "code_id",
          "debug_id",
          "architecture",
          "state",
          "qualification_version",
          "items",
          "next_cursor"
        ],
        "title": "CatalogPairOrigins",
        "type": "object"
      },
      "CatalogReviewEvidence": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "expected_version": {
            "title": "Expected Version",
            "type": "integer"
          },
          "pair_id": {
            "title": "Pair Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "schema_version": {
            "const": "catalog-provider-review-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "state": {
            "enum": [
              "active",
              "withdrawn"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "pair_id",
          "expected_version",
          "state",
          "reason",
          "reviewer",
          "evidence"
        ],
        "title": "CatalogReviewEvidence",
        "type": "object"
      },
      "CatalogReviewPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogReviewResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Version"
          }
        },
        "required": [
          "items",
          "next_version"
        ],
        "title": "CatalogReviewPage",
        "type": "object"
      },
      "CatalogReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "maxLength": 32000,
            "minLength": 1,
            "title": "Evidence",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "reviewer": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          },
          "state": {
            "enum": [
              "active",
              "withdrawn"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "state",
          "reason",
          "reviewer",
          "evidence",
          "idempotency_key"
        ],
        "title": "CatalogReviewRequest",
        "type": "object"
      },
      "CatalogReviewResponse": {
        "properties": {
          "evidence_sha256": {
            "title": "Evidence Sha256",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pair_id": {
            "title": "Pair Id",
            "type": "string"
          },
          "qualification_version": {
            "title": "Qualification Version",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "enum": [
              "active",
              "withdrawn"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pair_id",
          "qualification_version",
          "state",
          "reason",
          "evidence_sha256"
        ],
        "title": "CatalogReviewResponse",
        "type": "object"
      },
...
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorEnvelopeResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Anonymous trusted-intranet crash analysis control plane. There are intentionally no login, RBAC, or DELETE endpoints.",
    "title": "Crash-Cap API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/symbol-catalog/pairs/{pair_id}/origins": {
      "get": {
        "operationId": "get_pair_origins_api_v3_symbol_catalog_pairs__pair_id__origins_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPairOrigins"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Get Pair Origins"
      }
    },
    "/api/v3/symbol-catalog/pairs/{pair_id}/reviews": {
      "get": {
        "operationId": "list_pair_reviews_api_v3_symbol_catalog_pairs__pair_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Version"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "List Pair Reviews"
      },
      "post": {
        "operationId": "submit_pair_review_api_v3_symbol_catalog_pairs__pair_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Submit Pair Review"
      }
    },
    "/api/v3/symbol-catalog/pairs/{pair_id}/reviews/{review_id}/evidence": {
      "get": {
        "operationId": "get_review_evidence_api_v3_symbol_catalog_pairs__pair_id__reviews__review_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewEvidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Get Review Evidence"
      }
    },
...
  }
}